refactor -  auth module
